Crosses a major version boundary. Resolved 16 conflicts and fixed the iOS
build fallout, most of which produced no conflict at all because our fork had
removed code upstream later changed, or relied on transitive includes.

Conflict resolutions
- Accepted upstream's removal of discordrpc and the relocation of
  common/widgets/ -> widgets/; our edits there were iOS build workarounds for
  code upstream has since deleted.
- Took upstream's <SDL2/...> include form verbatim everywhere. Added a
  bin/iOS/sdl/SDL2 -> include symlink plus the parent dir on the include path
  so both forms resolve; this stops the whole class recurring on future merges.
- Kept our iOS hooks in the SDL platform layer (accelerometer hint, Vulkan
  force-on, I_GetCWD stack buffer, os_log fatal error) while taking upstream's
  changes around them.
- src/CMakeLists.txt: took upstream's abseil link additions.

iOS build fixes
- ios_start.mm: dropped 11 FBasicStartupScreen::Net* overrides, removed
  upstream when the netgame lobby moved to ZWidget; added the now-explicit
  basics.h include (st_start.h no longer pulls it in).
- iwadpicker_cocoa.mm: restored the TARGET_OS_IPHONE stub with the old
  signature. Upstream's sdl/i_system.cpp still calls it, so Apple + SDL backend
  does not link without it.
- ZMusic: skip CoreMIDI on iOS (no CoreAudio/HostTime.h); select the FluidSynth
  target by TARGET rather than the out-of-scope FLUIDSYNTH_FOUND.
- OpenAL: explicit empty iOS branch so find_package does not pull in the SDK's
  deprecated framework and the macOS-only AudioUnit/ApplicationServices.
- ZWidget: point it at the in-repo iOS SDL before add_subdirectory.
- name.cpp: raise -fconstexpr-steps for the constexpr PredefinedNames sort.
- NO_ASSERTS option to disable assert() in Debug builds (NDEBUG changes the
  VM_ARGS ABI, so it must be defined uniformly).
- Stage pk3s/soundfonts/fm_banks into the real .app via POST_BUILD; the generic
  path targets a macOS bundle layout that never ships on iOS.

Audio
- Take the static decoder path on iOS. ZMusic defaulted to dlopen'ing
  libmpg123.0.dylib / libsndfile.1.dylib, which cannot exist in an app bundle,
  so MP3, WAV and FLAC were silently unsupported. Link liblame.a and libopus.a,
  which libsndfile.a references once it is actually pulled into the link.

Other
- Fix alternate app icons: ALTERNATE_APPICON_NAMES must be space-separated, not
  semicolon-separated, or no CFBundleAlternateIcons is emitted and
  setAlternateIconName() silently does nothing.
- Show GenZD and UZDoom versions in the launcher footer.
- Add docs/ios-build.md documenting the architecture, build and known traps.

Known regression: iOS multiplayer. Upstream removed the FStartupScreen net API
entirely, so the IOS_ShowSystemModal/Bonjour hooks have no call site. Needs
rebuilding against the new ZWidget lobby.

+++ src/common/scripting/dap/Nodes/ValueStateNode.cpp
@@ -143,9 +143,9 @@ dap::Variable ValueStateNode::ToVariable(const VMValue &m_variable, PType *m_typ
 	{ // explicitly not TYPE_IntNotInt
 		int64_t val = TruncateVMValue(&m_variable, basic_type).i;
 		if (basic_type == BASIC_uint32 || basic_type == BASIC_uint16 || basic_type == BASIC_uint8){
-			variable.value = StringFormat("%lu", val);
+			variable.value = StringFormat("%llu", static_cast<uint64_t>(val));
 		} else {
-			variable.value = StringFormat("%ld", val);
+			variable.value = StringFormat("%lld", val);
 		}
 	}
 	else if (m_type->isFloat())
